Use "SFXContext" and "SFXContextObject" instead of "ctx" and "ctx" + "obj" to avoid confusion with Go contexts [+]
The SFX OpenURL API uses "ctx" as an XML namespace and makes use of a ContextObject to store metadata received by the SFX server:
https://developers.exlibrisgroup.com/sfx/apis/web_services/openurl/

In Go code, the `context.Context` type is used heavily in application code that deals with APIs and communication between servers (and processes).
The convention is to use the abbreviation "ctx" in variable names of type context.Context.

To avoid confusion in this code base, we will use "SFXContext" and "SFXContextObject" instead "Ctx"/"ctx" and "CtxObject"/"ctxObject"/"ctx_obj".
The exceptions to this are usages inside struct tags for definitions of XML and JSON structures and in XML templates, to preserve mapping with SFX API XML schemas.  These occurrences are easily distinguishable from usages in variable and function names, and are not likely to be ambiguous or confusing as typically one does not refer to `context.Context` inside strings and struct tags.