backend/Dockerfile.debug: restore `apk add build-base` [+]
While switching to golang:1.17 image allowed for `go install` of delve to complete without cgo-related errors, apparently the installed `dlv` can't be executed at runtime:

/app # ./util/dlv
/bin/sh: ./util/dlv: not found

I restored build-base and this cleared the error.  After finishing CI setup we can circle back and see if we can get this worked out, for a smaller and faster-built image.